Fix compilation errors: type casts, missing properties, and using statements
- PoliticalLadder.cs: Added explicit cast for long to int (BudgetSize)
- InfiniteUniverseManager.cs: Added System.Linq using for LINQ methods
- TournamentEconomyManager.cs: Added TransactionType alias
- ScandalConsequenceSystem.cs: Added explicit cast for float to int
- CharacterLibrary.cs: Added GetSavedCharacter method
- EventTemplateLibrary.cs: Fixed VariableMapping namespace (Templates not Translation)
- SaveManager.cs: Added Settings and Resources properties to LegacyGameSaveData
- Character.cs: Added CurrentTier and CurrentOfficeId properties
- WorldData.cs: Added TurnNumber property to World class
- NewsEventManager.cs: Changed lastFetchTime from DateTime to float

+++ Assets/Scripts/Gameplay/PoliticalLadder.cs
@@ -316,7 +316,7 @@ namespace ElectionEmpire.Gameplay
                     CanVetoLegislation = true,
                     CanDeclareEmergency = true,
                     CanControlBudget = true,
-                    BudgetSize = 5000000000,
+                    BudgetSize = (int)5000000000,
                     CanHireStaff = true,
                     MaxStaffSize = 50,
                     CanAppointJudges = true,
